Consolidate reused types
