--- packages/react/src/common/aspectRatio.ts
@@ -1,4 +1,0 @@
-export type AspectRatioProps = {
-  /** The aspect ratio to display media content in. */
-  aspectRatio?: 'x-tall' | 'tall' | 'square' | 'wide' | 'x-wide' | '2x-wide'
-}
