+++ packages/react/src/common/types.ts
@@ -1,3 +1,8 @@
+export type AspectRatioProps = {
+  /** The aspect ratio to display media content in. */
+  aspectRatio?: 'x-tall' | 'tall' | 'square' | 'wide' | 'x-wide' | '2x-wide'
+}
+
 export const crossAlignOptions = ['start', 'center', 'baseline', 'end'] as const
 export type CrossAlign = (typeof crossAlignOptions)[number]
