+++ storybook/src/components/Row/Row.stories.tsx
@@ -5,7 +5,7 @@
 
 import { Avatar, Heading, Link } from '@amsterdam/design-system-react'
 import { Row } from '@amsterdam/design-system-react/src'
-import { crossAlignOptions, mainAlignOptions } from '@amsterdam/design-system-react/src/common/layout'
+import { crossAlignOptions, mainAlignOptions } from '@amsterdam/design-system-react/src/common/types'
 import { Meta, StoryObj } from '@storybook/react'
 
 const ThreeItems = [
